L-01/requirements.txt
matplotlib==3.8.4
numpy==1.26.4
pillow==10.3.0
psutil==5.9.8
pytorch==2.3.1
scipy==1.13.0
tqdm==4.66.4
